add logic to create_namespace_env function
The basic idea is to have a router and a host in each network
namespace. The only way the hosts can communicate with each other
is through the openwrt routers connected through the veth devices
bridge to their tap interfaces. Such a setup could be used for much
more complex routing simulations in the future. A lot more logic
will need to be added eventually for create_namespace_env to be more
versatile, however, this should suffice for testing a typical
IPSec tunnel mode setup.

Signed-off-by: Alexandru Moise <00moses.alexander00@gmail.com>